+++ src/cpp/session/include/session/SessionModuleContext.hpp
@@ -872,7 +872,21 @@ void initializeConsoleCtrlHandler();
 
 bool isPythonReplActive();
 
+struct QuartoConfig
+{
+   QuartoConfig() : empty(true), installed(false), is_project(false) {}
+   bool empty;
+   bool installed;
+   bool is_project;
+   std::string project_type;
+   std::string project_output_dir;
+};
+
+QuartoConfig quartoConfig(bool refresh = false);
 bool isQuartoWebsiteDoc(const core::FilePath& filePath);
+
+std::vector<core::FilePath> ignoreContentDirs();
+bool isIgnoredContent(const core::FilePath& filePath, const std::vector<core::FilePath>& ignoreDirs);
 
 std::string getActiveLanguage();
 core::Error adaptToLanguage(const std::string& language);