+++ src/cpp/session/modules/SessionPythonEnvironments.R
@@ -128,7 +128,6 @@
       }
    }
    
-   # nothing found
    ""
 })