+++ src/gwt/src/org/rstudio/studio/client/quarto/model/QuartoConfig.java
@@ -20,4 +20,7 @@ import jsinterop.annotations.JsType;
 public class QuartoConfig
 {
    public boolean installed;
+   public boolean is_project;
+   public String project_type;
+   public String project_output_dir;
 }