+++ src/gwt/src/org/rstudio/studio/client/shiny/ShinyApplication.java
@@ -17,6 +17,7 @@ package org.rstudio.studio.client.shiny;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.google.gwt.event.shared.HandlerRegistration;
 import org.rstudio.core.client.BrowseCap;
 import org.rstudio.core.client.Size;
 import org.rstudio.core.client.StringUtil;
@@ -30,6 +31,7 @@ import org.rstudio.studio.client.application.ApplicationInterrupt.InterruptHandl
 import org.rstudio.studio.client.application.events.EventBus;
 import org.rstudio.studio.client.application.events.RestartStatusEvent;
 import org.rstudio.studio.client.common.GlobalDisplay;
+import org.rstudio.studio.client.common.Value;
 import org.rstudio.studio.client.common.dependencies.DependencyManager;
 import org.rstudio.studio.client.common.satellite.SatelliteManager;
 import org.rstudio.studio.client.common.satellite.events.WindowClosedEvent;
@@ -47,6 +49,10 @@ import org.rstudio.studio.client.workbench.prefs.model.UserPrefs;
 import org.rstudio.studio.client.workbench.views.console.events.ConsoleBusyEvent;
 import org.rstudio.studio.client.workbench.views.console.events.SendToConsoleEvent;
 import org.rstudio.studio.client.workbench.views.environment.events.DebugModeChangedEvent;
+import org.rstudio.studio.client.workbench.views.jobs.events.JobUpdatedEvent;
+import org.rstudio.studio.client.workbench.views.jobs.model.Job;
+import org.rstudio.studio.client.workbench.views.jobs.model.JobConstants;
+import org.rstudio.studio.client.workbench.views.jobs.model.JobManager;
 import org.rstudio.studio.client.workbench.views.jobs.model.JobsServerOperations;
 import org.rstudio.studio.client.workbench.views.viewer.events.ViewerClearedEvent;
 
@@ -77,6 +83,7 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
                            JobsServerOperations jobsServer,
                            GlobalDisplay display,
                            DependencyManager dependencyManager,
+                           JobManager jobManager,
                            ApplicationInterrupt interrupt)
    {
       eventBus_ = eventBus;
@@ -89,6 +96,7 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
       isBusy_ = false;
       currentViewType_ = UserPrefs.SHINY_VIEWER_TYPE_NONE;
       dependencyManager_ = dependencyManager;
+      jobManager_ = jobManager;
       interrupt_ = interrupt;
       params_ = new ArrayList<>();
       
@@ -132,6 +140,38 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
       }
       else if (event.getParams().getState() == ShinyApplicationParams.STATE_STOPPED)
       {
+         // If this Shiny application was running in the background, we're responsible for ending the job.
+         if (event.getParams().isBackgroundApp())
+         {
+            if (StringUtil.equals(stoppingId_, event.getParams().getId()))
+            {
+               // This application ended because we asked it to; no need to stop the job (it's already in the process
+               // of stopping)
+               stoppingId_ = null;
+            }
+            else
+            {
+               for (ShinyApplicationParams params: params_)
+               {
+                  if (StringUtil.equals(params.getId(), event.getParams().getId()))
+                  {
+                     // If we got here, the Shiny application that was stopped corresponds to a background job. If the
+                     // job is still running, stop it.
+                     Job job = jobManager_.getJob(params.getId());
+                     if (job != null)
+                     {
+                        if (job.state == JobConstants.STATE_RUNNING)
+                        {
+                           params.setState(ShinyApplicationParams.STATE_STOPPING);
+                           stopShinyBackgroundJob(params.getId(), null);
+                        }
+                     }
+
+                     break;
+                  }
+               }
+            }
+         }
          params_.removeIf(params -> params.getId() == event.getParams().getId());
       }
    }
@@ -292,6 +332,43 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
             });
             return;
          }
+         else if (StringUtil.equals(destination, BACKGROUND_APP) && params.isBackgroundApp())
+         {
+            // This app is running in the background; we need to restart it. This is a multi-step process
+            // since it is not possible to replay a currently running job; we must first stop it and wait
+            // for it to terminate, and only then initiate a replay.
+
+            // Stop the background job -- this returns immediately (it's an async call)
+            params.setState(ShinyApplicationParams.STATE_STOPPING);
+            stopShinyBackgroundJob(params.getId(), () ->
+            {
+               // Now wait for a JobUpdatedEvent that tells us that the job has actually stopped.
+               final Value<HandlerRegistration> reg = new Value<>(null);
+               reg.setValue(eventBus_.addHandler(JobUpdatedEvent.TYPE, evt ->
+               {
+                  if (StringUtil.equals(params.getId(), evt.getData().job.id) &&
+                     evt.getData().job.state != JobConstants.STATE_RUNNING)
+                  {
+                     // The job is now stopped; replay it
+                     jobsServer_.executeJobAction(params.getId(), JobConstants.ACTION_REPLAY,
+                        new VoidServerRequestCallback()
+                        {
+                           @Override
+                           public void onError(ServerError error)
+                           {
+                              display_.showErrorMessage("Failed to reload",
+                                 "Could not reload the Shiny application.\n\n" +
+                                    error.getMessage());
+                           }
+                        });
+                     // Clean up event handler
+                     reg.getValue().removeHandler();
+                  }
+               }));
+            });
+
+            return;
+         }
       }
       
       // Nothing else running, start this app.
@@ -374,31 +451,47 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
       {
          if (appState.getId() == ShinyApplicationParams.ID_FOREGROUND)
          {
-            // Foreground apps: interrupt R itself
-            if (commands_.interruptR().isEnabled()) 
+            // If this app is running in the foreground, stop it now. Background jobs will be stopped when we fire the
+            // status event (below).
+            if (commands_.interruptR().isEnabled())
                commands_.interruptR().execute();
-         }
-         else
-         {
-            // Background apps: stop the associated job
-            jobsServer_.executeJobAction(appState.getId(), "stop", 
-                  new VoidServerRequestCallback()
-            {
-               @Override
-               public void onError(ServerError error)
-               {
-                  display_.showErrorMessage("Failed to Stop", 
-                        "Could not stop the Shiny application.\n\n" + 
-                        error.getMessage());
-               }
-            });
          }
          appState.setState(ShinyApplicationParams.STATE_STOPPED);
       }
       eventBus_.fireEvent(new ShinyApplicationStatusEvent(
             (ShinyApplicationParams) params.cast()));
    }
-   
+
+   /**
+    * Stops the Shiny background job associated with the given ID
+    */
+   private void stopShinyBackgroundJob(String id, Command onSuccess)
+   {
+      // Save the ID of the job we are stopping; this allows us to distinguish between jobs that we stopped ourselves
+      // and jobs that were ended for other reasons.
+      stoppingId_ = id;
+      jobsServer_.executeJobAction(id, JobConstants.ACTION_STOP,
+         new VoidServerRequestCallback()
+         {
+            @Override
+            public void onSuccess()
+            {
+               if (onSuccess != null)
+               {
+                  onSuccess.execute();
+               }
+            }
+
+            @Override
+            public void onError(ServerError error)
+            {
+               display_.showErrorMessage("Failed to Stop",
+                  "Could not stop the Shiny application.\n\n" +
+                     error.getMessage());
+            }
+         });
+   }
+
    private final native void exportShinyAppClosedCallback()/*-{
       var registry = this;     
       $wnd.notifyShinyAppClosed = $entry(
@@ -560,6 +653,7 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
    
    private final EventBus eventBus_;
    private final SatelliteManager satelliteManager_;
+   private final JobManager jobManager_;
    private final DependencyManager dependencyManager_;
    private final Commands commands_;
    private final Provider<UserPrefs> pPrefs_;
@@ -574,7 +668,8 @@ public class ShinyApplication implements ShinyApplicationStatusEvent.Handler,
    private boolean stopOnNextClose_ = true;
    private String satelliteClosePath_ = null;
    private String currentViewType_;
-   
+   private String stoppingId_ = null;
+
    public static final String FOREGROUND_APP = "foreground";
    public static final String BACKGROUND_APP = "background";
 }