+++ src/node/desktop/src/main/utils.ts
@@ -105,3 +105,33 @@ export function removeStaleOptionsLockfile(): void {
 
   fs.unlinkSync(lockFilePath);
 }
+
+export function rsessionExeName(): string {
+  return process.platform === 'win32' ? 'rsession.exe' : 'rsession';
+}
+
+/**
+ * @returns Paths to config file, rsession, and desktop scripts.
+ */
+export function findComponents(): [FilePath, FilePath, FilePath] {
+
+  // determine paths to config file, rsession, and desktop scripts
+  let confPath: FilePath = new FilePath();
+  let sessionPath: FilePath = new FilePath();
+
+  const binRoot = new FilePath(app.getAppPath());
+  if (app.isPackaged) {
+    // confPath is intentionally left empty for a package build
+    sessionPath = binRoot.completePath(`bin/${rsessionExeName()}`);
+  } else {
+    const buildRootEnv = getenv('RSTUDIO_CPP_BUILD_OUTPUT');
+    if (!buildRootEnv) {
+      throw Error('RSTUDIO_CPP_BUILD_OUTPUT env var must contain ' +
+        'path where src/cpp was built (dev config).');
+    }
+    const buildRoot = new FilePath(buildRootEnv);
+    confPath = buildRoot.completePath('conf/rdesktop-dev.conf');
+    sessionPath = buildRoot.completePath(`session/${rsessionExeName()}`);
+  }
+  return [confPath, sessionPath, new FilePath(app.getAppPath())];
+}